chore: refacto services, add specialization
